apify-actors/Developer_Tools/ip-whois-inspector/requirements.txt
apify < 4.0.0
ipwhois
dnspython
geoip2
requests